renamed common-components folder
