+++ app/pages/supplier/dashboard/template.njk
@@ -1,7 +1,7 @@
 {% extends 'includes/layout.njk' %}
 {% from 'components/button/macro.njk' import button %}
 {% from './components/dashboard-sectionGroup.njk' import dashboardSectionGroup %}
-{% from '../../../common-components/error-summary/macro.njk' import errorSummary %}
+{% from '../../../common/components/error-summary/macro.njk' import errorSummary %}
 
 {% block body %}
 <div data-test-id="dashboard" class="nhsuk-grid-row">
