+++ app/pages/supplier/section/template.njk
@@ -3,7 +3,7 @@
 {% from 'components/button/macro.njk' import button %}
 {% from './components/section-question.njk' import sectionQuestion %}
 {% from './components/multi-question.njk' import multiQuestion %}
-{% from '../../../common-components/error-summary/macro.njk' import errorSummary %}
+{% from '../../../common/components/error-summary/macro.njk' import errorSummary %}
 
 {% block body %}
 <div class="nhsuk-grid-row">
